Вернул приложение users назад, убрал весь свой код из других приложений, вернул настройки для users
