+++ api_yamdb/api/serializers.py
@@ -1,5 +1,4 @@
 from django.contrib.auth import get_user_model
-from rest_framework import serializers
 from rest_framework.serializers import ModelSerializer
 from rest_framework.validators import UniqueTogetherValidator
 
@@ -27,25 +26,3 @@ class CommentSerializer(ModelSerializer):
     class Meta:
         model = Comment
         fields = '__all__'
-
-
-# Сериализаторы для авторизации и работы с пользователями: начало
-
-class UserRegistrationSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ('username', 'email')
-
-    def validate_username(self, value):
-        if value == 'me':
-            raise serializers.ValidationError(
-                {'username': ('Недопустимое значение.',)}
-            )
-        return value
-
-
-class ConfirmationSerializer(serializers.Serializer):
-    username = serializers.CharField(max_length=150)
-    confirmation_code = serializers.CharField(max_length=36)
-
-# Сериализаторы для авторизации и работы с пользователями: конец
