+++ api_yamdb/reviews/admin.py
@@ -1,13 +1,6 @@
 from django.contrib import admin
-from django.contrib.auth.admin import UserAdmin
 
-from .models import Review, Comment, ApiUser
-
-UserAdmin.fieldsets += (
-    ('Extra Fields', {'fields': ('bio','role')}),
-)
-UserAdmin.list_display += ('role',)
-admin.site.register(ApiUser, UserAdmin)
+from .models import Review, Comment
 
 
 @admin.register(Review)
