+++ api_yamdb/users/serializers.py
@@ -1,0 +1,27 @@
+"""users/serializers.py
+
+Serializers для приложения users.
+"""
+
+from django.contrib.auth import get_user_model
+from rest_framework import serializers
+
+User = get_user_model()
+
+
+class UserRegistrationSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ('username', 'email')
+
+    def validate_username(self, value):
+        if value == 'me':
+            raise serializers.ValidationError(
+                {'username': ('Недопустимое значение.',)}
+            )
+        return value
+
+
+class ConfirmationSerializer(serializers.Serializer):
+    username = serializers.CharField(max_length=150)
+    confirmation_code = serializers.CharField(max_length=36)
